remove fn RPITIT
Originally the RPITIT was intended to enable parallel sort/filtering by
applying different Send/Sync constraint to returned closures, however,
it doesn't seem to implement the desired behavior in foreseeable future.

Without the behavior, the RPIIT is just a source of confusion which
simply makes the derived closure less readable.