Fix width of placeholder for sort indicator.
Notes:
It doesn't make any sense to use `ui.monospace(" ");` since the sort indicator is 'n' characters long because it contains an indicator and a number and that number can be 1-n columns which may result in a number that is 1 or more characters long. e.g. sort indicator length for the 10th sorted column would be 3. So using only a single character here makes the width of the column wider but never wide enough.